only use time_shared_mutex from core module, minimize changes
