+++ plugins/fclrave/CMakeLists.txt
@@ -71,7 +71,10 @@ if( FCL_FOUND )
     add_library(fclrave SHARED fclrave.cpp fclcollision.h fclstatistics.h fclspace.h plugindefs.h)
     target_link_libraries(fclrave libopenrave ${FCL_LIBRARIES})
     target_link_libraries(fclrave PRIVATE boost_assertion_failed)
-    set_target_properties(fclrave PROPERTIES COMPILE_FLAGS "${PLUGIN_COMPILE_FLAGS} ${FCL_CFLAGS_OTHER}" LINK_FLAGS "${PLUGIN_LINK_FLAGS} ${FCL_LDFLAGS}")
+    if( CMAKE_COMPILER_IS_GNUCC OR CMAKE_COMPILER_IS_GNUCXX OR COMPILER_IS_CLANG)
+      add_definitions("-std=c++11")
+    endif()
+    set_target_properties(fclrave PROPERTIES COMPILE_FLAGS "${PLUGIN_COMPILE_FLAGS} ${FCL_CFLAGS_OTHER} -std=c++11" LINK_FLAGS "${PLUGIN_LINK_FLAGS} ${FCL_LDFLAGS}")
     install(TARGETS fclrave DESTINATION ${OPENRAVE_PLUGINS_INSTALL_DIR} COMPONENT ${COMPONENT_PREFIX}plugin-fclrave)
   else()
     message(STATUS "Could not find FCL. Please install FCL (https://github.com/flexible-collision-library/fcl)")
