+++ python/databases/inversekinematics.py
@@ -1086,6 +1086,7 @@ class InverseKinematicsModel(DatabaseGenerator):
             if compiler.compiler_type == 'unix':
                 compile_flags.append('-O3')
                 compile_flags.append('-fPIC')
+                compile_flags.append('-std=c++11')
         return compiler,compile_flags
     
     @staticmethod
